Create Online Quiz App Part
Prepare to move the database from local to cloud, we got we will use firebase to do that. This part is the sign in / sign up part of the Online Quiz App Part.
And in this part we got a bug, we  didn't get a code error, but when we try to run this program,  the android terminal say that to stop the program.

Co-Authored-By: Anders kuang <kuanjianchen@users.noreply.github.com>